requirements.txt
asgiref==3.5.2
Django==4.1.2
sqlparse==0.4.3
django-bootstrap4
djangorestframework==3.14.0
django-cors-headers
django_extensions
requests